[Tests] Update checks
